requirements.txt
anyio==4.3.0
APScheduler==3.6.3
asgiref==3.8.1
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Django==5.0.4
django-environ==0.11.2
djangorestframework==3.15.1
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.51.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.8.4
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-telegram-bot==13.7
pytz==2024.1
requests==2.31.0
scikit-image==0.23.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tifffile==2024.4.18
tornado==6.4
tzlocal==5.2
urllib3==2.2.1
